# docs/swagger.yaml
definitions:
  dto.CreateSubscriptionRequest:
    properties:
      end_date:
        type: string
      price:
        minimum: 1
        type: integer
      service_name:
        type: string
      start_date:
        type: string
      user_id:
        type: string
    required:
    - price
    - service_name
    - start_date
    - user_id
    type: object
  dto.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  dto.SubscriptionResponse:
    properties:
      end_date:
        type: string
      id:
        type: integer
      price:
        type: integer
      service_name:
        type: string
      start_date:
        type: string
      user_id:
        type: string
    type: object
  dto.UpdateSubscriptionRequest:
    properties:
      end_date:
        type: string
      price:
        minimum: 1
        type: integer
      service_name:
        type: string
      start_date:
        type: string
      user_id:
        type: string
    type: object
info:
  contact: {}
paths:
  /subscriptions:
    get:
      description: Получить список всех подписок
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.SubscriptionResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Получить все подписки
      tags:
      - subscriptions
    post:
      consumes:
      - application/json
      description: Создать новую запись о подписке пользователя
      parameters:
      - description: Данные подписки
        in: body
        name: subscription
        required: true
        schema:
          $ref: '#/definitions/dto.CreateSubscriptionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.SubscriptionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Создать подписку
      tags:
      - subscriptions
  /subscriptions/{id}:
    delete:
      description: Удалить подписку по ID
      parameters:
      - description: ID подписки
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Удалить подписку
      tags:
      - subscriptions
    get:
      description: Получить запись подписки по ID
      parameters:
      - description: ID подписки
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SubscriptionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Получить подписку по ID
      tags:
      - subscriptions
    put:
      consumes:
      - application/json
      description: Обновить запись подписки по ID
      parameters:
      - description: ID подписки
        in: path
        name: id
        required: true
        type: integer
      - description: Обновленные данные подписки
        in: body
        name: subscription
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateSubscriptionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SubscriptionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Обновить подписку
      tags:
      - subscriptions
  /subscriptions/total:
    get:
      description: Подсчитывает общую стоимость подписок за период с фильтрацией по
        user_id и service_name
      parameters:
      - description: UUID пользователя
        in: query
        name: user_id
        required: true
        type: string
      - description: Название сервиса
        in: query
        name: service_name
        type: string
      - description: Дата начала периода (dd-MM-YYYY)
        in: query
        name: from
        type: string
      - description: Дата конца периода (dd-MM-YYYY)
        in: query
        name: to
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: integer
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Получить суммарную стоимость подписок
      tags:
      - subscriptions
swagger: "2.0"